medium-xrender: remove custom {gcontext,drawable}-picture
These functions were used by the font renderer. This is a step towards uniform
rendering model using pictures.

A new function is introduced called medium-source-picture. Currently it works
only for uniform colors - this is not a regression, because the font renderer
took gcontext-foreground color disregarding of a design - using +deep-pink+ as
a fallback is just making this issue apparent.